# internal/http/openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VBPClaw Gateway API",
    "description": "PostgreSQL multi-tenant AI agent gateway with WebSocket RPC + HTTP API.\n\n## Authentication\n\nAll endpoints require a Bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <gateway-token-or-api-key>\n```\n\nYou can use either the **gateway token** (grants admin access) or an **API key** created via the API Keys endpoints (grants scoped access).\n\nIf no token is configured on the server, authentication is disabled (backward compatibility).\n\n## Common Headers\n\n| Header | Description |\n|--------|-------------|\n| `X-VBPClaw-User-Id` | External user ID for multi-tenant context |\n| `X-VBPClaw-Agent-Id` | Target agent ID (alternative to model prefix) |\n| `X-VBPClaw-Tenant-Id` | Tenant scope — UUID or slug (gateway token / cross-tenant API keys) |\n| `Accept-Language` | Locale for error messages (`en`, `vi`, `zh`) |\n\n## WebSocket Protocol\n\nConnect via `POST /ws` (upgrade). Protocol v3 uses frame types: `req`, `res`, `event`.\nFirst request must be `connect` with `{\"token\": \"...\", \"user_id\": \"...\", \"locale\": \"en\"}`.",
    "version": "0.2.0",
    "contact": {
      "name": "VBPClaw",
      "url": "https://github.com/kadiesnguyen/vbpclaw"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/skills": {
      "get": {
        "tags": ["Skills"],
        "summary": "List all skills",
        "responses": { "200": { "description": "Skill list" } }
      },
      "post": {
        "tags": ["Skills"],
        "summary": "Upload skill",
        "requestBody": { "content": { "multipart/form-data": {} } },
        "responses": { "201": { "description": "Skill uploaded" } }
      }
    },
    "/v1/skills/{id}": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Skill details" } }
      },
      "put": {
        "tags": ["Skills"],
        "summary": "Update skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Skill updated" } }
      },
      "delete": {
        "tags": ["Skills"],
        "summary": "Delete skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Skill deleted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gateway token or API key"
      }
    },
...
  }
}